# src/backend/api/docs/swagger.yaml
basePath: /api
definitions:
  decimal.NullDecimal:
    properties:
      decimal:
        type: number
      valid:
        type: boolean
    type: object
...
  model.Connector:
    properties:
      connector_specific_config:
        $ref: '#/definitions/model.JSONMap'
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      docs:
        items:
          $ref: '#/definitions/model.Document'
        type: array
      docs_map:
        additionalProperties:
          $ref: '#/definitions/model.Document'
        type: object
      id:
        type: number
      last_successful_analysis:
        $ref: '#/definitions/pg.NullTime'
      last_update:
        $ref: '#/definitions/pg.NullTime'
      name:
        type: string
      refresh_freq:
        type: integer
      source:
        $ref: '#/definitions/model.SourceType'
      status:
        type: string
      tenant_id:
        $ref: '#/definitions/uuid.NullUUID'
      total_docs_indexed:
        type: integer
      user_id:
        type: string
    type: object
  model.Document:
    properties:
      analyzed:
        type: boolean
      chunking_session:
        $ref: '#/definitions/uuid.NullUUID'
      connector_id:
        type: number
      creation_date:
        type: string
      id:
        type: number
      last_update:
        $ref: '#/definitions/pg.NullTime'
      parent_id:
        $ref: '#/definitions/decimal.NullDecimal'
      signature:
        type: string
      source_id:
        type: string
      url:
        type: string
    type: object
...
  model.EmbeddingModel:
    properties:
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      id:
        type: number
      is_active:
        type: boolean
      last_update:
        $ref: '#/definitions/pg.NullTime'
      model_dim:
        type: integer
      model_id:
        type: string
      model_name:
        type: string
      tenant_id:
        type: string
      url:
        type: string
    type: object
  model.JSONMap:
    additionalProperties: true
    type: object
...
  model.SourceType:
    enum:
    - ingestion_api
    - slack
    - web
    - google_drive
    - gmail
    - requesttracker
    - github
    - gitlab
    - guru
    - bookstack
    - confluence
    - slab
    - jira
    - productboard
    - file
    - notion
    - zulip
    - linear
    - hubspot
    - document360
    - gong
    - google_sites
    - zendesk
    - loopio
    - sharepoint
    - one-drive
    - msteams
    - youtube
    type: string
    x-enum-varnames:
    - SourceTypeIngestionApi
    - SourceTypeSlack
    - SourceTypeWEB
    - SourceTypeGoogleDrive
    - SourceTypeGMAIL
    - SourceTypeRequesttracker
    - SourceTypeGithub
    - SourceTypeGitlab
    - SourceTypeGuru
    - SourceTypeBookstack
    - SourceTypeConfluence
    - SourceTypeSlab
    - SourceTypeJira
    - SourceTypeProductboard
    - SourceTypeFile
    - SourceTypeNotion
    - SourceTypeZulip
    - SourceTypeLinear
    - SourceTypeHubspot
    - SourceTypeDocument360
    - SourceTypeGong
    - SourceTypeGoogleSites
    - SourceTypeZendesk
    - SourceTypeLoopio
    - SourceTypeSharepoint
    - SourceTypeOneDrive
    - SourceTypeMsTeams
    - SourceTypeYoutube
  model.SourceTypeDescription:
    properties:
      id:
        $ref: '#/definitions/model.SourceType'
      isImplemented:
        type: boolean
      name:
        type: string
    type: object
...
  parameters.CreateConnectorParam:
    properties:
      connector_specific_config:
        $ref: '#/definitions/model.JSONMap'
      disabled:
        type: boolean
      name:
        type: string
      refresh_freq:
        type: integer
      shared:
        type: boolean
      source:
        type: string
    type: object
...
  parameters.EmbeddingModelParam:
    properties:
      is_active:
        type: boolean
      model_dim:
        type: integer
      model_id:
        type: string
      model_name:
        type: string
      url:
        type: string
    type: object
...
  parameters.UpdateConnectorParam:
    properties:
      connector_specific_config:
        $ref: '#/definitions/model.JSONMap'
      name:
        type: string
      refresh_freq:
        type: integer
      shared:
        type: boolean
      status:
        type: string
    type: object
  pg.NullTime:
    properties:
      time.Time:
        type: string
    type: object
  uuid.NullUUID:
    properties:
      uuid:
        type: string
      valid:
        description: Valid is true if UUID is not NULL
        type: boolean
    type: object
info:
  contact:
    name: API Support
  description: This is Cognix Golang API Documentation
  title: Cognix API
  version: "1.0"
paths:
...
  /manage/connector:
    get:
      description: return list of allowed connectors
      operationId: connectors_get_all
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Connector'
            type: array
      security:
      - ApiKeyAuth: []
      summary: return list of allowed connectors
      tags:
      - Connectors
  /manage/connector/:
    post:
      description: creates connector
      operationId: connectors_create
      parameters:
      - description: connector create parameter
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/parameters.CreateConnectorParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Connector'
      security:
      - ApiKeyAuth: []
      summary: creates connector
      tags:
      - Connectors
  /manage/connector/{id}:
    get:
      description: return list of allowed connectors
      operationId: connectors_get_by_id
      parameters:
      - description: connector id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Connector'
      security:
      - ApiKeyAuth: []
      summary: return list of allowed connectors
      tags:
      - Connectors
    put:
      description: updates connector
      operationId: connectors_update
      parameters:
      - description: connector id
        in: path
        name: id
        required: true
        type: integer
      - description: connector update parameter
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/parameters.UpdateConnectorParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Connector'
      security:
      - ApiKeyAuth: []
      summary: updates connector
      tags:
      - Connectors
  /manage/connector/{id}/{action}:
    post:
      description: delete or restore connector
      operationId: Connectors_delete_restore
      parameters:
      - description: Connectors id
        in: path
        name: id
        required: true
        type: integer
      - description: 'action : delete | restore '
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Connector'
      security:
      - ApiKeyAuth: []
      summary: delete or restore connector
      tags:
      - Connectors
  /manage/connector/source_types:
    get:
      description: return list of source types
      operationId: connectors_get_source_types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SourceTypeDescription'
            type: array
      security:
      - ApiKeyAuth: []
      summary: return list of source types
      tags:
      - Connectors
  /manage/embedding_models:
    get:
      description: return list of embedding models
      operationId: embedding_model_get_all
      parameters:
      - description: true for include deleted embedding models
        in: query
        name: archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.EmbeddingModel'
            type: array
      security:
      - ApiKeyAuth: []
      summary: return list of embedding models
      tags:
      - EmbeddingModel
    post:
      description: creates embedding models
      operationId: embedding_model_create
      parameters:
      - description: embedding model parameter
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/parameters.EmbeddingModelParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.EmbeddingModel'
      security:
      - ApiKeyAuth: []
      summary: creates embedding models
      tags:
      - EmbeddingModel
  /manage/embedding_models/{id}:
    get:
      description: return embedding model by id
      operationId: embedding_model_get_by_id
      parameters:
      - description: embedding model id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EmbeddingModel'
      security:
      - ApiKeyAuth: []
      summary: return embedding model by id
      tags:
      - EmbeddingModel
    put:
      description: updates embedding model
      operationId: embedding_model_update
      parameters:
      - description: embedding model id
        in: path
        name: id
        required: true
        type: integer
      - description: embedding model parameter
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/parameters.EmbeddingModelParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EmbeddingModel'
      security:
      - ApiKeyAuth: []
      summary: updates embedding model
      tags:
      - EmbeddingModel
  /manage/embedding_models/{id}/{action}:
    post:
      description: delete or restore embedding model
      operationId: embedding_model_delete
      parameters:
      - description: embedding model id
        in: path
        name: id
        required: true
        type: integer
      - description: 'action : delete | restore '
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EmbeddingModel'
      security:
      - ApiKeyAuth: []
      summary: delete or restore embedding model
      tags:
      - EmbeddingModel
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"